test(guardrails): split test_apply_guardrail into themed files under 500 LOC
Break the 899-line test_apply_guardrail.py into three focused files: text-side
actions stay in test_apply_guardrail.py, tool-call / tool-definition / legacy
functions[] scanning moves to test_apply_guardrail_tools.py, and fail-open/closed
plus missing-secrets and helpers move to test_apply_guardrail_failmodes.py. No
test bodies changed; every alice file is now under 500 added LOC.
